fix(ui): log out using endpoint to clear jwt
